# backend/app/routers/local_admin.py
from __future__ import annotations

import json
from datetime import UTC, datetime
from typing import Any, Literal

from fastapi import APIRouter, Depends, HTTPException, Request, status
from pydantic import BaseModel, Field

from app.config import get_settings
from app.schemas import InteractiveTravelState
from app.services.interactive_travel_finale_service import (
    DEFAULT_DIRECTION,
    build_interactive_travel_story,
    compile_interactive_travel_video_prompt,
    generate_interactive_travel_finale_video,
    list_interactive_travel_finale_attempts,
    list_interactive_travel_finales,
    read_interactive_travel_finale,
    save_interactive_travel_finale,
)
from app.services.local_admin_publish import (
    AdminPublishError,
    get_admin_publish_job,
    read_admin_manifest_from_server,
    start_admin_publish,
    sync_admin_files_from_server,
)
from app.services.local_admin_store import read_admin_manifest, save_admin_files

LOCAL_HOSTS = {"127.0.0.1", "::1", "localhost", "testclient"}


class AdminFileUpdate(BaseModel):
    id: str = Field(min_length=1, max_length=80)
    content: str


class AdminSaveRequest(BaseModel):
    files: list[AdminFileUpdate] = Field(min_length=1, max_length=16)


class AdminPublishRequest(BaseModel):
    files: list[AdminFileUpdate] = Field(default_factory=list, max_length=16)
    message: str | None = Field(default=None, max_length=180)


class TravelFinaleImportRequest(BaseModel):
    travel: InteractiveTravelState
    telegramId: int = 62943754
    username: str | None = Field(default=None, max_length=64)
    firstName: str | None = Field(default=None, max_length=64)


class TravelFinalePromptRequest(BaseModel):
    direction: str = Field(min_length=1, max_length=8000)


class TravelFinaleGenerateRequest(BaseModel):
    prompt: str = Field(min_length=1, max_length=16000)
    referenceBaseUrl: str = Field(min_length=1, max_length=1000)


AdminSource = Literal["local", "production"]


class AdminSpeechFileResponse(BaseModel):
    id: str
    label: str
    path: str
    format: Literal["json"]
    description: str
    exists: bool
    sizeBytes: int
    updatedAt: str | None
    summary: dict[str, Any]
    content: str


class AdminSyncResponse(BaseModel):
    status: str
    message: str
    serverCommit: str | None
    updatedAt: str


class AdminDeployResponse(BaseModel):
    enabled: bool
    message: str


class AdminSpeechManifestResponse(BaseModel):
    generatedAt: str
    mode: AdminSource
    files: list[AdminSpeechFileResponse]
    sync: AdminSyncResponse
    deploy: AdminDeployResponse


class AdminSavedFileResponse(BaseModel):
    id: str
    path: str
    backupPath: str | None
    sizeBytes: int


class AdminSaveResponse(BaseModel):
    saved: bool
    updatedAt: str
    errors: dict[str, str]
    files: list[AdminSavedFileResponse]


class AdminPublishLogResponse(BaseModel):
    at: str
    level: str
    message: str


class AdminPublishJobResponse(BaseModel):
    id: str
    status: Literal["pending", "running", "succeeded", "failed"]
    createdAt: str
    startedAt: str | None
    finishedAt: str | None
    logs: list[AdminPublishLogResponse]
    error: str | None
    errorCode: str | None
    savedFiles: list[AdminSavedFileResponse]
    commitSha: str | None


def _now_iso() -> str:
    return datetime.now(UTC).isoformat().replace("+00:00", "Z")


def _sync_warning_result(exc: AdminPublishError) -> dict[str, Any]:
    return {
        "status": "local_dirty",
        "message": (
            "Есть незадеплоенные изменения. "
            "Форма показывает локальные data-файлы; нажми Deploy, когда они готовы."
        ),
        "serverCommit": None,
        "updatedAt": _now_iso(),
    }


def require_local_admin(request: Request) -> None:
    settings = get_settings()
    host = request.client.host if request.client else ""
    if settings.allow_dev_tma_auth and host in LOCAL_HOSTS:
        return
    raise HTTPException(
        status_code=status.HTTP_404_NOT_FOUND,
        detail={"code": "LOCAL_ADMIN_DISABLED", "message": "Local admin is disabled."},
    )


router = APIRouter(
    prefix="/api/admin",
    tags=["local-admin"],
    dependencies=[Depends(require_local_admin)],
)


@router.get("/travel-finales")
def travel_finale_list() -> dict[str, Any]:
    return {"items": list_interactive_travel_finales()}


@router.post("/travel-finales/import")
def travel_finale_import(payload: TravelFinaleImportRequest) -> dict[str, Any]:
    return save_interactive_travel_finale(
        payload.travel,
        telegram_id=payload.telegramId,
        username=payload.username,
        first_name=payload.firstName,
    )


@router.get("/travel-finales/{travel_id}")
def travel_finale_detail(travel_id: str) -> dict[str, Any]:
    try:
        snapshot = read_interactive_travel_finale(travel_id)
    except (FileNotFoundError, ValueError, json.JSONDecodeError) as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Travel finale not found",
        ) from exc
    travel = InteractiveTravelState.model_validate(snapshot["travel"])
    return {
        **snapshot,
        "story": build_interactive_travel_story(travel),
        "defaultDirection": DEFAULT_DIRECTION,
        "attempts": list_interactive_travel_finale_attempts(travel_id),
    }


@router.post("/travel-finales/{travel_id}/prompt")
def travel_finale_prompt(travel_id: str, payload: TravelFinalePromptRequest) -> dict[str, str]:
    try:
        snapshot = read_interactive_travel_finale(travel_id)
    except (FileNotFoundError, ValueError, json.JSONDecodeError) as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Travel finale not found",
        ) from exc
    prompt = compile_interactive_travel_video_prompt(
        InteractiveTravelState.model_validate(snapshot["travel"]),
        direction=payload.direction,
    )
    return {"prompt": prompt}


@router.post("/travel-finales/{travel_id}/generate")
def travel_finale_generate(
    travel_id: str,
    payload: TravelFinaleGenerateRequest,
) -> dict[str, Any]:
    try:
        snapshot = read_interactive_travel_finale(travel_id)
    except (FileNotFoundError, ValueError, json.JSONDecodeError) as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Travel finale not found",
        ) from exc
    return generate_interactive_travel_finale_video(
        InteractiveTravelState.model_validate(snapshot["travel"]),
        prompt=payload.prompt,
        reference_base_url=payload.referenceBaseUrl,
    )


@router.get("/speech", response_model=AdminSpeechManifestResponse)
def speech_admin_manifest(source: AdminSource = "local") -> dict[str, Any]:
    settings = get_settings()
    deploy_enabled = bool(getattr(settings, "admin_publish_enabled", False))
    deploy_message = (
        ("Deploy отправит data-файлы в GitHub и применит их на Hetzner.")
        if deploy_enabled
        else ("Deploy отключен. Задай ADMIN_PUBLISH_ENABLED=true и SSH-настройки.")
    )
    try:
        if source == "production":
            return read_admin_manifest_from_server(
                settings,
                deploy_enabled=deploy_enabled,
                deploy_message=deploy_message,
            )
        sync_result = None
        if getattr(settings, "admin_sync_from_server_enabled", False):
            try:
                sync_result = sync_admin_files_from_server(settings)
            except AdminPublishError as exc:
                if exc.code != "ADMIN_SYNC_LOCAL_DIRTY":
                    raise
                sync_result = _sync_warning_result(exc)
        return read_admin_manifest(
            deploy_enabled=deploy_enabled,
            deploy_message=deploy_message,
            sync_result=sync_result,
        )
    except AdminPublishError as exc:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail={"code": exc.code, "message": exc.message},
        ) from exc


@router.put("/speech", response_model=AdminSaveResponse)
def save_speech_admin(payload: AdminSaveRequest, source: AdminSource = "local") -> dict[str, Any]:
    if source == "production":
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail={
                "code": "ADMIN_PRODUCTION_DIRECT_SAVE_DISABLED",
                "message": "Production меняется только через publish/deploy.",
            },
        )
    result = save_admin_files([item.model_dump() for item in payload.files])
    if not result["saved"]:
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail=result,
        )
    return result


@router.post("/speech/publish", response_model=AdminPublishJobResponse)
def publish_speech_admin(payload: AdminPublishRequest) -> dict[str, Any]:
    try:
        return start_admin_publish(
            files=[item.model_dump() for item in payload.files],
            settings=get_settings(),
            commit_message=payload.message,
        )
    except AdminPublishError as exc:
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail={"code": exc.code, "message": exc.message},
        ) from exc


@router.get("/speech/publish/{job_id}", response_model=AdminPublishJobResponse)
def speech_admin_publish_job(job_id: str) -> dict[str, Any]:
    result = get_admin_publish_job(job_id)
    if result is None:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail={"code": "ADMIN_PUBLISH_JOB_NOT_FOUND", "message": "Publish job not found."},
        )
    return result