# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AdminDeployResponse": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "message"
        ],
        "title": "AdminDeployResponse",
        "type": "object"
      },
      "AdminFileUpdate": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content"
        ],
        "title": "AdminFileUpdate",
        "type": "object"
      },
      "AdminPublishJobResponse": {
        "properties": {
          "commitSha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commitsha"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/AdminPublishLogResponse"
            },
            "title": "Logs",
            "type": "array"
          },
          "savedFiles": {
            "items": {
              "$ref": "#/components/schemas/AdminSavedFileResponse"
            },
            "title": "Savedfiles",
            "type": "array"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "startedAt",
          "finishedAt",
          "logs",
          "error",
          "errorCode",
          "savedFiles",
          "commitSha"
        ],
        "title": "AdminPublishJobResponse",
        "type": "object"
      },
      "AdminPublishLogResponse": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "at",
          "level",
          "message"
        ],
        "title": "AdminPublishLogResponse",
        "type": "object"
      },
      "AdminPublishRequest": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/AdminFileUpdate"
            },
            "maxItems": 16,
            "title": "Files",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 180,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "title": "AdminPublishRequest",
        "type": "object"
      },
      "AdminSaveRequest": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/AdminFileUpdate"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "AdminSaveRequest",
        "type": "object"
      },
      "AdminSaveResponse": {
        "properties": {
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Errors",
            "type": "object"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/AdminSavedFileResponse"
            },
            "title": "Files",
            "type": "array"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "saved",
          "updatedAt",
          "errors",
          "files"
        ],
        "title": "AdminSaveResponse",
        "type": "object"
      },
      "AdminSavedFileResponse": {
        "properties": {
          "backupPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backuppath"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "path",
          "backupPath",
          "sizeBytes"
        ],
        "title": "AdminSavedFileResponse",
        "type": "object"
      },
      "AdminSpeechFileResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "format": {
            "const": "json",
            "title": "Format",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "required": [
          "id",
          "label",
          "path",
          "format",
          "description",
          "exists",
          "sizeBytes",
          "updatedAt",
          "summary",
          "content"
        ],
        "title": "AdminSpeechFileResponse",
        "type": "object"
      },
      "AdminSpeechManifestResponse": {
        "properties": {
          "deploy": {
            "$ref": "#/components/schemas/AdminDeployResponse"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/AdminSpeechFileResponse"
            },
            "title": "Files",
            "type": "array"
          },
          "generatedAt": {
            "title": "Generatedat",
            "type": "string"
          },
          "mode": {
            "enum": [
              "local",
              "production"
            ],
            "title": "Mode",
            "type": "string"
          },
          "sync": {
            "$ref": "#/components/schemas/AdminSyncResponse"
          }
        },
        "required": [
          "generatedAt",
          "mode",
          "files",
          "sync",
          "deploy"
        ],
        "title": "AdminSpeechManifestResponse",
        "type": "object"
      },
      "AdminSyncResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "serverCommit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servercommit"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "serverCommit",
          "updatedAt"
        ],
        "title": "AdminSyncResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InteractiveTravelState": {
        "properties": {
          "arcPlan": {
            "additionalProperties": {
              "maxLength": 500,
              "type": "string"
            },
            "maxProperties": 32,
            "propertyNames": {
              "maxLength": 80,
              "minLength": 1
            },
            "title": "Arcplan",
            "type": "object"
          },
          "completed": {
            "default": false,
            "title": "Completed",
            "type": "boolean"
          },
          "destination": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "introReaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InteractiveTravelIntroReaction"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcomeValence": {
            "anyOf": [
              {
                "enum": [
                  "positive",
                  "negative"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcomevalence"
          },
          "overallTitle": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Overalltitle",
            "type": "string"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/InteractiveTravelPart"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Parts",
            "type": "array"
          },
          "statImpact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InteractiveTravelStatImpact"
              },
              {
                "type": "null"
              }
            ]
          },
          "travelId": {
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^interactive-travel-[A-Za-z0-9_-]+$",
            "title": "Travelid",
            "type": "string"
          }
        },
        "required": [
          "travelId",
          "generatedAt",
          "destination",
          "overallTitle",
          "arcPlan",
          "parts"
        ],
        "title": "InteractiveTravelState",
        "type": "object"
      },
...
      "TravelFinaleGenerateRequest": {
        "properties": {
          "prompt": {
            "maxLength": 16000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "referenceBaseUrl": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Referencebaseurl",
            "type": "string"
          }
        },
        "required": [
          "prompt",
          "referenceBaseUrl"
        ],
        "title": "TravelFinaleGenerateRequest",
        "type": "object"
      },
      "TravelFinaleImportRequest": {
        "properties": {
          "firstName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "telegramId": {
            "default": 62943754,
            "title": "Telegramid",
            "type": "integer"
          },
          "travel": {
            "$ref": "#/components/schemas/InteractiveTravelState"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "travel"
        ],
        "title": "TravelFinaleImportRequest",
        "type": "object"
      },
      "TravelFinalePromptRequest": {
        "properties": {
          "direction": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Direction",
            "type": "string"
          }
        },
        "required": [
          "direction"
        ],
        "title": "TravelFinalePromptRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Tamagotchi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/speech": {
      "get": {
        "operationId": "speech_admin_manifest_api_admin_speech_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "local",
              "enum": [
                "local",
                "production"
              ],
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSpeechManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speech Admin Manifest",
        "tags": [
          "local-admin"
        ]
      },
      "put": {
        "operationId": "save_speech_admin_api_admin_speech_put",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "local",
              "enum": [
                "local",
                "production"
              ],
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSaveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Speech Admin",
        "tags": [
          "local-admin"
        ]
      }
    },
    "/api/admin/speech/publish": {
      "post": {
        "operationId": "publish_speech_admin_api_admin_speech_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPublishJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Speech Admin",
        "tags": [
          "local-admin"
        ]
      }
    },
    "/api/admin/speech/publish/{job_id}": {
      "get": {
        "operationId": "speech_admin_publish_job_api_admin_speech_publish__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPublishJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speech Admin Publish Job",
        "tags": [
          "local-admin"
        ]
      }
    },
...
    "/api/admin/travel-finales/import": {
      "post": {
        "operationId": "travel_finale_import_api_admin_travel_finales_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelFinaleImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Travel Finale Import Api Admin Travel Finales Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Travel Finale Import",
        "tags": [
          "local-admin"
        ]
      }
    },
    "/api/admin/travel-finales/{travel_id}": {
      "get": {
        "operationId": "travel_finale_detail_api_admin_travel_finales__travel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "travel_id",
            "required": true,
            "schema": {
              "title": "Travel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Travel Finale Detail Api Admin Travel Finales  Travel Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Travel Finale Detail",
        "tags": [
          "local-admin"
        ]
      }
    },
    "/api/admin/travel-finales/{travel_id}/generate": {
      "post": {
        "operationId": "travel_finale_generate_api_admin_travel_finales__travel_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "travel_id",
            "required": true,
            "schema": {
              "title": "Travel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelFinaleGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Travel Finale Generate Api Admin Travel Finales  Travel Id  Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Travel Finale Generate",
        "tags": [
          "local-admin"
        ]
      }
    },
    "/api/admin/travel-finales/{travel_id}/prompt": {
      "post": {
        "operationId": "travel_finale_prompt_api_admin_travel_finales__travel_id__prompt_post",
        "parameters": [
          {
            "in": "path",
            "name": "travel_id",
            "required": true,
            "schema": {
              "title": "Travel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelFinalePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Travel Finale Prompt Api Admin Travel Finales  Travel Id  Prompt Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Travel Finale Prompt",
        "tags": [
          "local-admin"
        ]
      }
    },
...
  }
}